Blink animation for cursors
