+++ edit-frontend/styles/edit.scss
@@ -239,8 +239,26 @@ body.theme-column {
         margin-left: -1px;
         margin-right: -1px;
         content: '\200c' !important; // zero width non-joiner Unicode code point
-        animation: fadein 0.3s;
         pointer-events: none;
+        animation: blink-animation 2s linear infinite;
+    }
+
+    @keyframes blink-animation {
+        0% {
+            opacity: 1.0;
+        }
+        43% {
+            opacity: 1.0;
+        }
+        50% {
+            opacity: 0.0;
+        }
+        93% {
+            opacity: 0.0;
+        }
+        100% {
+            opacity: 1.0;
+        }
     }
 
     div[data-tag="caret"][data-focus="false"] {
@@ -251,6 +269,7 @@ body.theme-column {
         border-left: 2px $color_curcaret solid;
         z-index: 100;
         position: relative;
+        animation: blink-animation 1s linear infinite;
     }
 
     // TODO the overlapping dashed cursors isn't working well
