+++ oatie/src/core/doc.rs
@@ -4,26 +4,11 @@ use serde::{
     Deserialize,
     Serialize,
 };
-use std::collections::HashMap;
 use crate::style::OpaqueStyleMap;
 
 // Re-exports
 pub use super::place::*;
 pub use crate::string::*;
-
-#[derive(Clone, PartialEq, Debug, Serialize, Deserialize)]
-pub enum Attrs {
-    Header(u8),
-    Text,
-    Code,
-    Html,
-    ListItem,
-    Rule,
-    Caret {
-        client_id: String,
-        focus: bool,
-    },
-}
 
 pub type DocSpan = Vec<DocElement>;
 pub type DelSpan = Vec<DelElement>;
@@ -32,14 +17,9 @@ pub type CurSpan = Vec<CurElement>;
 
 pub type Op = (DelSpan, AddSpan);
 
-/// Returns true if the style map can be omitted.
-fn is_stylemap_empty(map: &OpaqueStyleMap) -> bool {
-    map.iter().count() == 0
-}
-
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub enum DocElement {
-    DocChars(DocString, #[serde(default, skip_serializing_if = "is_stylemap_empty")] OpaqueStyleMap),
+    DocChars(DocString, #[serde(default, skip_serializing_if = "OpaqueStyleMap::is_empty")] OpaqueStyleMap),
     DocGroup(Attrs, DocSpan),
 }
 
@@ -72,7 +52,7 @@ pub use self::DelElement::*;
 pub enum AddElement {
     AddSkip(usize),
     AddWithGroup(AddSpan),
-    AddChars(DocString, #[serde(default, skip_serializing_if = "is_stylemap_empty")] OpaqueStyleMap),
+    AddChars(DocString, #[serde(default, skip_serializing_if = "OpaqueStyleMap::is_empty")] OpaqueStyleMap),
     AddGroup(Attrs, AddSpan),
     AddStyles(usize, StyleMap),
 }
